pyproject.toml
[tool.poetry]
name = "multi-agent-concierge-main"
version = "0.0.1"
description = "It is scalable LLM system that could work with the images"
authors = ["Your Name <you@example.com>"]
license = "0.0.2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index-core = "^0.11.19"
llama-index-llms-openai = "^0.2.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
llama-index-llms-anthropic = "^0.2.0"
llama-index-agent-openai = "^0.3.0"
llama-index-utils-workflow = "^0.2.2"
redis = {extras = ["hiredis"], version = "^5.2.1"}
pinecone = "^6.0.1"
psycopg2 = "^2.9.10"
transformers = "^4.49.0"
torch = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
